Fix the meltdown's blast powers, and drop the lever's hand-rolled quad
CI caught one mistake made five times: the blast scale comes from the config as
a double, every blast power is a float, and float * double is a double that
javac will not narrow back on its own. Narrowed once at the source rather than
at each call site. tools/syntax-check.sh cannot see this - it is a type error,
not a structural one, which is exactly the limit its own output advertises.

While in there, the lever's indicator lamp was drawn as a hand-built quad with
raw addVertex/setUv/setLight calls - by some distance the riskiest code in this
feature, and unverifiable against the decompiled sources because they do not
ship VertexConsumer. It is now a second ModelPart with two textures, one green
and one red. Same result on screen, no raw vertex work, and it uses only
ModelPart.render, which the turbine and the windmill have both been exercising
for months.

+++ citiesinlife/src/main/java/com/branciho/citiesinlife/nuclear/Meltdown.java
@@ -114,7 +114,9 @@ public final class Meltdown {
     private static boolean advance(ServerLevel level, Structure plant, ReactorState state) {
         ReactorSurvey survey = ReactorSurvey.of(level, plant.min(), plant.max());
         int t = state.meltdownTick++;
-        double scale = CitiesInLifeConfig.nuclearBlastScale();
+        // Narrowed here rather than at each call site: every blast power is a float, and
+        // float * double is a double that javac will not silently narrow back.
+        float scale = (float) CitiesInLifeConfig.nuclearBlastScale();
 
         if (t == 0) {
             state.output = 0;
